component 분리 및 구조 수정
