+++ src/components/CommunityMain/index.tsx
@@ -1,3 +1,4 @@
 export { default as CommunityMain } from './CommunityMain';
+export { default as CommunityList } from './CommunityList';
 export { default as CommunityItem } from './CommunityItem';
 export { default as AddContentButton } from './AddContentButton'
